requirements.txt
SQLAlchemy~=2.0.37
python-dotenv~=1.0.1
openai~=1.61.0
aiogram~=3.17.0
asyncpg